PcRemote (Android-Java)
An Android based mobile application, by which we can connect a mobile
device to a computer and control the
computer remotely. The app has options to shutdown, Restart, Volume
up/down/mute, Lock screen, play media,
create file and delete files on the computer. Apart of these, the app
can be used to give any system command to be
executed by the computer. This whole process is facilitated by a java
program pre-installed on the computer.